requirements.txt
flask~=3.1.1
flask-sqlalchemy
flask-login~=0.6.3
flask-wtf
python-dotenv~=1.1.1
openai~=1.95.0
python-docx~=1.2.0
fpdf2~=2.8.3
markdown2~=2.5.3
email-validator
wtforms~=3.2.1
Werkzeug~=3.1.3
Markdown~=3.8.2
fpdf~=1.7.2
click~=8.2.1
dotenv~=0.9.9